IdeaHub/apps/profile/models.py
from django.db import models
from django.contrib.auth.models import User
import uuid


class Profile(models.Model):
    id = models.UUIDField(
        primary_key=True,
        default=uuid.uuid4,
        editable=False,
        unique=True
    )
    user = models.OneToOneField(User, on_delete=models.CASCADE)


class VerificationCode(models.Model):
    id = models.UUIDField(
        primary_key=True,
        default=uuid.uuid4,
        editable=False,
        unique=True
    )
    code = models.CharField(max_length=250)
    user = models.OneToOneField(User, on_delete=models.CASCADE)
